tests/wpt/web-platform-tests/workers/Worker_cross_origin_security_err.htm

<!DOCTYPE html>
<title> Worker cross-origin URL </title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
async_test(function(t) {
  try {
    var w = new Worker("ftp://example.org/support/WorkerBasic.js");
    w.onerror = t.step_func_done(function(e) {
      assert_true(e instanceof ErrorEvent);
    });
  } catch (e) {
    t.step_func_done(function(e) { assert_true(true); });
  }
});
</script>
